requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
twilio==9.3.0
python-dotenv==1.0.1
python-multipart==0.0.12
google-generativeai==0.8.3
